requirements.txt
streamlit
pandas
requests
google-cloud-aiplatform
google-cloud-bigquery
